Added a host manager.
The intended application is a simple process which operates on the default port and will open up child hosts on demand. This is primarily a convenience for development, where you often want a persistent and detached process which operates on an open/random port, but is still discoverable via env introspection.